Switch to stable Rust, replacing maud with ructe.
Maud is the sole reason this project uses nightly, and as nightly does not reliably provide rustfmt or clippy, it'd be nice to migrate off of it. Replacing maud with ructe allows stable-only features.

Ructe itself is a bit less well integrated, requiring more build infrastructure and a bit more code in the template dispatcher. This change could use further elaboration, but I'm satisfied with it as a proof of concept.